src/cli/main.py
"""
CLI Interface for Dai Assistant.
Entry point for command-line operations.
"""
import click
from rich.console import Console

console = Console()

@click.group()
@click.version_option()
def cli():
    """Dai Assistant CLI - Orchestrate collaborative AI agents."""
    pass

@cli.command()
@click.option("--project-name", "-p", help="Name of the project.")
def init(project_name):
    """Initialize a new project with default agent configuration."""
    console.print(f"[bold green]Initializing project:[/bold green] {project_name}")
    # TODO: Implementation for project initialization

@cli.command()
@click.option("--name", "-n", required=True, help="Name of the agent.")
@click.option("--role", "-r", required=True, help="Role of the agent.")
@click.option("--backstory", "-b", help="Backstory for the agent.")
@click.option("--output-format", "-o", help="Expected output format for the agent.")
def define_agent(name, role, backstory, output_format):
    """Define a new agent with specific role and characteristics."""
    console.print(f"[bold green]Defining agent:[/bold green] {name} as {role}")
    # TODO: Implementation for agent definition

@cli.command()
@click.option("--agents-config", "-a", required=True, help="Path to agents configuration file.")
@click.option("--project-name", "-p", required=True, help="Name of the project.")
def run(agents_config, project_name):
    """Run the multi-agent orchestration with specified configuration."""
    console.print(f"[bold green]Running project:[/bold green] {project_name}")
    console.print(f"[bold blue]Using agents configuration:[/bold blue] {agents_config}")
    # TODO: Implementation for running the orchestration

@cli.command()
@click.option("--config-path", "-c", required=True, help="Path to configuration file.")
def load_config(config_path):
    """Load an existing configuration file."""
    console.print(f"[bold green]Loading configuration:[/bold green] {config_path}")
    # TODO: Implementation for loading configuration

if __name__ == "__main__":
    cli()
